Refactor XML specific app cast maker items to a subclass
